+++ src/Contracts/StorageInterface.php
@@ -38,5 +38,5 @@ interface StorageInterface
 
     public function ensureSpatialTableExists(string $name): void;
 
-    public function getIndexedTerms(?string $indexName = null, int $minFrequency = 2, int $limit = 10000): array;
+    public function getIndexedTerms(?string $indexName = null, int $minFrequency = 1, int $limit = 10000): array;
 }